rename wasm_module to module
