+++ hermes/bin/src/packaging/module/author_payload.rs
