+++ hermes/bin/src/packaging/module/manifest.rs
